Tweaks the section about 3rd party info
As this was written, it was a bit weird because I assume some (most?) of your systems have PII, financial info, and proprietary info. Encountering it could be fairly normal -- for example, I might see my own PII, but that doesn't mean I should stop doing my vulnerability test.

The important thing is *does the vulnerability* allow access to *more* of it.